Moved some common functions to util class
